test: ensure processes are ready for tracing in tst.manypids.tst
This test depends on preloading a lot of libraries which can cause the
task startup to be slower than expected.  As a result, dtrace may fail
to enable probes.  This is typically the case if the preload libraries
and executable are not in cache yet.

We explicitly invoke /usr/bin/sleep right before starting the tasks to
be traced to prime the cache.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>