Consolidate like keys for @[mod()], @[sym()], etc.
Certain actions like mod() and sym() take address arguments and
report string values.  These values can be used as aggregation
keys.

DTrace happens to pass the raw addresses from producer to consumer,
for the consumer to convert to strings and combine like keys for
aggregation.  The reduction of common values had been omitted in the
DTrace-to-Linux port.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>