test: Update tst.aggmod_full2.sh
Remove the xfail now that duplicate mod() keys have been reduced to
unique values.

Also, update the test to handle the large number of addresses passed
from producer to consumer.  In DTrace, @[mod(addr)] is handled by
passing individual addresses to the consumer, which is responsible
for reducing to unique mod() values.  This test marches through all
symbols found in /proc/kall[mod]syms and might therefore pass close
to 200k addresses.  In legacy DTrace, this was handled with aggrate
and throttling the trigger program.  These techniques do not work in
the DTrace on Linux port, which keeps building the BPF aggregation
maps.  So, remove the throttle and expand aggsize.  Also, reduce
strsize to alleviate the pressure on aggsize.

Finally, ignore the problematic __builtin__ftrace module.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>