libproc: fix mispositioned Pwait()
rd_new() contains a Pwait() outside a Ptrace()/Puntrace() pair.
This usually works because rd_new is usually called for processes
that are already Ptrace()-stopped, but if it's not, this fails
horribly (ptrace() returns -ECHILD, Ptrace() misfires and
erroneously concludes the process is dead).

This Pwait() is completely unnecessary: we do a Ptrace() shortly after it
(before we do anything else that depends on tracee process state), and
Ptrace() does a Pwait() itself as part of its normal operation anyway
(syncing up our idea of the process state with its actual state).

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>