libproc: fix Ptrace() error returns
Ptrace() is defined to return a positive number (really zero) on success or
a negative number on error.  Those relatively few call sites that can do
anything useful when Ptrace() fails all do negative-number checks...

... so it's a shame that it also sometimes returns, uh, PS_DEAD, which is,
uh, 4, and sometimes returns an errno, which is a positive number.

Negate the latter and make a dead process return -ECHILD.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>