libproc: detect dead process faster when updating maps
Scanning maps and lmids is really quite slow: it involves parsing text files
out of /proc and digging around in the dynamic linker and even doing
searches of victim process memory to detect changes in internal dynamic
linker data structures.  So, in Pupdate_syms(), after we do that, check if
the process has died before we run off and update all the per-file mappings,
which will all fail noisily if the process has died.

(Just a cosmetic improvement, but it can reduce the amount of debugging
output by hundreds of megabytes sometimes.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>