proc: improve armouring against self-grabs
The DTrace victim process-monitoring loop (used by -c in particular, but also
by u*() actions) has some protection to keep DTrace from trying to ptrace()
itself: processes whose PID is the same as DTrace's, or which is already
being traced by DTrace, or which is a system daemon, do not get traced.

But this is not quite good enough. DTrace is a multithreaded process, but we
check the tracer pid (which is a kernel task ID derived from
/proc/$pid/status, i.e. a *thread* ID) against getpid() (which is a POSIX
pid, i.e. a *task group* ID).  We should dig out the task group ID of the
process being traced and compare *that* against our own PID.  (This still
does not stop two distinct dtraces trying to trace each other, but in that
case they shouldn't deadlock: each should sometimes let the other run and
they'll both make some sort of stuttering forward progress.)

To armour things a bit further, we shouldn't try to maintain a Pwait()ing
process-monitoring loop for anything libproc has already said it cannot
ptrace().  It doesn't matter whether we *asked* for it to be noninvasively
traced in the first place: what matters is purely whether we got a ptrace()
grab on it, because that's what controls whether we need to do the whole
waitpid() signal-redirection dance to it.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>