test: fix two return tests to not cause trigger buffer overruns
test/triggers/bogus-ioctl.c is rather odd.  It was originally written for
test/unittest/io/tst.fds.d.  It uses bogus ioctl()s to trigger two things:

 - an infloop at the start with first arg -1, which the script responds to
   by raise()ing SIGUSR1 to get it going again via a signal handler and a
   siglongjmp() (you could use -x evaltime=main for the same purpose, but
   from the dates I believe this test predates that feature)
 - a loop at the end which passes the fd number as the first arg, which
   the script responds to by printing out that fd

A lot of other scripts have grown up that rely on it to just spray out
bogus ioctls forever, but alas two have grown up that assume that they
can get it going via hitting it with SIGUSR1 whenever a bogus ioctl()
is received, regardless of its arg values or anything.  This works fine
the first time, but the second time around it acts as a goto back to the
setjmp(), increments n too far, triggers a buffer overrun on the fds
array in bogus-ioctl and then bogus-ioctl hits an assertion failure and
dies.

This went unnoticed before now because it often happens that the assertion
failure and coredump takes long enough that runtest.sh has already tested
for a coredump by the time it happens, but it's still a bug.  Fix trivial,
we don't even need to look at the arg values: only raise(SIGUSR1) at most
once.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>