sdt: retrieve probe arguments from CTF data (if possible)
We were determining the probe argument types based on data from the
TRACEFS/events/<group>/<probe>/format file.  If the CTF type for the
trace event data struct can be found, we can use that instead.

If no such CTF data is available, we fall back to the EVENTFS based
approach.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>