lex: fix scan buffer memory leak
When compiling a string, the dt_compile() function calls yy_scan_string()
on the string, which replaces the curerent buffer with a buffer based on
the given string.  The buffer being replaced does not get freed, causing
a memory leak.  Popping the buffer state frees the existing buffer.  It
will leave the lexer buffer state NULL, which allows the string-based
buffer to be installed without leaks.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>