bpf, btf: store BTF dict id for kernel modules
BPF requires a file descriptor (provided by the kernel) to be associated
with BTF type ids for loadble kernel modules.  This file descriptor can
only be obtained from the kernel based on a BTF dict id for the module,
and these ids are assigned dynamically at runtime when modules are
loaded.

After kernel modules are initialized for DTrace, the corresponding BTF
dict ids are associated with their modules.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>