dlibs: fix access to `major_names data
On systems that lack CTF data, all kernel variable have a default type of
uintptr_t.  That means that an explicit &-operator needs to be applied
is the kernel variable is e.g. the address of an array because it will not
be recognized as a ref-type.

Fix this for `major_names access in the io translators.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>