test/triggers: ensure triggers use a reasonable fd limit
Various triggers have a loop to close all possible fds, and they are
loop from 0 to rlim_max.  On some systems, that was such a high limit
that the loop was taking too long, causing tests to time out.  Set
the soft limit to 1024, and loop up to that limit (rlim_cur).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>