test: adjust err.unloaded_var.x to match test
Commit d9158b26 ("test: adjust err.unloaded_var.d again.") adjusted the
test but failed to adjust the matching .x file.  Along with changing
the kernel module name (and location), we can also drop the testing for
the CTF archive since it is not necessary.  In addition, we allow for
either gfs2.ko or gfs2.ko.xz to handle compressed kernel modules.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>