Change probes from having lists of clauses to lists of stmts
Each stmt has a clause and a probe description.  Traditionally,
we have added clauses to probes.  Further, we have generated an
enabled probe ID for each probe/clause combination, and the
consumer has used the EPID to determine the PRID as well as the
data descriptor for the clause.

In this patch, change from probes having clauses associated
with them to having stmts.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>