test: Allow duplicate usym/umod/uaddr if for different pids
In 83da884cbdc5 ("Preface usym/umod/uaddr with pid"), a bug was fixed
in which addresses in the same module (or function) might be mapped to
multiple agg entries.  This fix helped the associated tests run much
more successfully.  Nonetheless, tests would sometimes still fail.

Another problem is that the tests themselves were overly narrow.  It
is fine for a module (or function) to appear multiple times in the
aggregation output... if those entries correspond to different pids.

Further, odd behaviors can result for some of the processes running on
a system.

Change the tests to add a "pid" agg key.  Filter on only a few, select
pids.  Distinguish agg entries by pid.

There are still occasional time outs observed with these tests,
presumably because the tick-2s probe is not firing (when profile-1234hz
is running).

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>