test: Embellish the USDT multiprovider test
Add a probe to two providers, each probe with the same name but
different arguments.

Do so in a new, separate test since the older test used to trigger
some crashes, and we would like to continue to catch such problems.

Add -v to these probe-list tests to check attributes and arguments.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>