test: USDT probes with multiple providers and shared libs
A simple executable links in two shared libraries.  There is
a different USDT provider for the executable and for each of
the two libraries.  Each provider has probes named the same
as for the other providers, but the probes differ in arguments.

Run dtrace on the executable.  The D script first dumps the
target pid for later post-processing.  All USDT probes are
traced and their argument values tested.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>