Move proc lock to where we actually find a USDT process
The function dt_pid_create_usdt_probes_proc() creates USDT probes
for a specific process.  It is called in one of two ways:

    dt_proc_scan(dtp, dpr) ->
    dt_pid_create_probes_module(dtp, dpr)

        The process has been locked and we are updating its
        USDT probes.

    dt_pid_create_usdt_probes(pdp, dtp, pcb)

        Here, we look for any pids that might have USDT probes,
        calling the pid-specific function for each candidate.

One problem is that the first code path assumes the process is locked.
This means the second code path has to lock processes even before we
know if it has any USDT probes we care about.

Change dt_pid_create_usdt_probes_proc() so the caller can specify the
process in either one of two ways:

    by pid (implying the process has not been locked)

    by dpr (implying the process has been locked)

In the first case, the function will lock the process, but only if
USDT probes have been found.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>