symtab: add support for 'traceable' flag and add dt_symbol_by_name_next()
The traceable flag can be set on symbols to indicate that the symbol can
be traced using kernel facilities.

The traceable flag can be set on a symtab to indicte that its symbols
have traceable flags associated with them.

When the same symbol exists in multiple modules, dt_symbol_by_name()
returns reference to the first one.  The dt_symbol_by_name_next()
function is used to walk through the rest of the list (if any).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>