Fix trampoline use of BPF stack for scratch
The trampoline starts by setting %r9 = %fp + -DCTX_SIZE -- that is,
%r9 = %fp - 88.  Then, it fills the BPF stack from %fp-88 to %fp with
the various dctx-> pointers.  Calls to clauses will reset %fp.

The proc provider's trampoline was using the BPF stack for scratch
space for the exit and signal-handle probes.  Specifically, it used
%fp + DT_STK_SPILL(0), which overwrites a dctx-> pointer.

Switch to DT_TRAMP_SP_SLOT(0), which is intended for this purpose.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>