procfs: implement execargs for pr_psargs translator support
Implement a new execargs built-in variable to provide the ps-style
argument string for the current task.  It is used in the psinfo
translator for the pr_psargs member.

Due to BPF limitations, it is only possible to retrieve this information
for the current task.  The data is stored in the task's address space,
and the only BPF helper that can access data in another task's address
space is not allowed to be used in non-sleepable programs (such as
tracing programs).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>