trace: fix char-array handling
The special handling of strings in the consumer's trace implementation
did not correctly account for char-arrays that do not necessarily are
intended to hold string data.  The trace action is documented to print
data as an ASCII string if it can be represented as such, using the
heuristics implemented in dt_print_bytes().

The implementation for trace on both producer and consumer side was
using the data alignment as a marker to determine whether data should
be printed as a string (alignment == 1) or not.  The real alignment
of the data should be used.

The dtrace_diftype struct has its dtdt_pad member renamed as dtdt_align.

Added several new test cases to cover various possibilities.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>