test: Account for pid:::entry ustack() being correct
The pid:::entry uprobe fires so early in the function preamble
that the frame pointer is not yet set and the caller is not (yet)
correctly identified.  In Linux 6.11, x86-specific heuristics
address this problem.

Post process results from this test to accommodate both cases --
missing caller and not missing caller.

Orabug: 37459289
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>