test: Check tid value
We were checking the built-in variable tid simply by testing
that we could print it and its value was not -1.

Add a test that confirms the value is actually correct;
compare to C output of gettid().

In line with other similar tests, also check for the profile
provider.

While we're at it, check the pid value and the pthread_t value
returned via pthread_create().

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>