Clean up sched provider trampoline FIXMEs
The sched provider trampoline for enqueue and dequeue probes had
pending FIXMEs for providing a cpuinfo_t* for the cpu associated
with the run queue.  Implement the missing code.

Since the cpu associated with the run queue might be different from
the cpu where we are running, it becomes necessary to access the
cpuinfo for some random cpu.  With Linux 5.18, there is a BPF
helper function map_lookup_percpu_elem() that allows such lookups
on per-cpu arrays.  To support older kernels, however, we change
the cpuinfo BPF map from per-cpu to global.  Also, it is a hash
table rather than an array in case cpus are not numbered consecutively.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>