test: Get cpc expected branches and instructions counts from perf
For a number of the cpc tests, we get expected counts from perf.
For branches and instructions, however, we can determine the
expected counts more directly since there is one branch and a
fixed number of instructions per iteration.  Thus, we can derive
an expected cpc counts simply by knowing the number of iterations.

For some compilers, however, there is apparently some loop unrolling
even at low, default levels of optimization.

So, revert to the perf count to estimate the expected cpc count
even for the branches and instructions tests.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>