test: Skip pid-0 tests on oversubscribed systems
A number of tests check "tick-n /pid==0/" probes.  The problem
with this is that a tick-n probe runs on a specific CPU.  If that
CPU is fully subscribed, then pid 0 (swapper) will not run.  Thus,
the test will take a long time, only to time out.

Change these tests to use profile-n instead of tick-n probes,
improving chances that the test probe will fire on a less subscribed
CPU.

Therefore, also change the .r.p post-processing file so that it uses
only one output line (in case two CPUs manage to write output).

Finally, add skip files in case pid 0 does not fire on any CPU.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>