test: Mimic dtrace arithmetic more closely for avg/stddev
The multicpus test checks that data from multiple CPUs is aggregated
properly.  Operations like avg() and stddev() require division.
DTrace uses integer division, while awk does not.

Change the awk check to truncate intermediate results after division.

In practice, intermediate results are typically integer values anyhow.
So the test has generally passed.  Non-integer values could arise if,
for example, CPUs are not numbered consecutively.  More typically,
there may be a problem that the profile probe is not firing on every
expected CPU.  So a test failure probably was a sign of a problem,
but that's a different problem, one beyond the scope of this test.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>